fix: move validateEventsQuery to lib to fix Next.js route export type error
Next.js only allows HTTP method exports (GET, POST, etc.) from route files.
'validateEventsQuery' being exported from route.ts caused a build failure:
  Type error: 'validateEventsQuery' is not a valid Route export field.

Fix: extract to src/lib/events/validate.ts and import from both route.ts
and route.test.ts.
